backend/requirements/base.txt
pytz==2021.1
psycopg2-binary==2.9.1
Pillow==8.3.2
# Django
# ------------------------------------------------------------------------------
django==3.2.7
django-environ==0.7.0
django-treebeard==4.5.1
django-cors-headers==3.8.0
whitenoise==5.3.0
# Django rest framework
# ------------------------------------------------------------------------------
djangorestframework==3.12.4

## test
pytest==6.2.5
factory-boy==3.2.0
pytest-django==4.4.0

# Code quality
# ------------------------------------------------------------------------------
flake8==3.9.2
mypy==0.910
black==21.9b0
isort==5.9.3
